Use ST_Intersects function to exclude
Hi Andy! I am not sure if this needs editing. Both methods work (and apparently the one you use is much more simple), but since the question mentions the ST_Intersects function, I think maybe the one right here is what you expect of us to do. But anyway, I think it's fine if you leave it as the "!=" format. I just do a pull request in case you didn't make it this way on purpose.